requirements.txt
PyQt6-Fluent-Widgets>=1.5.5